Migrate fjs/bnf/module.f.ts to .f.mjs, split types into types.ts
Splits the type-level API (TerminalRange, Sequence, Variant, DataRule,
LazyRule, Rule, RangeVariant, None, Option, Repeat0Plus, Repeat1Plus,
Join1Plus, Join0Plus, Repeat) into a sibling types.ts, and converts
the runtime implementation (rangeEncode, oneEncode, eof, fullRange,
unicodeRange, unicodeMax, rangeDecode, toSequence, str, set, range,
remove, not, notSet, none, option, repeat0Plus, repeat1Plus,
join1Plus, join0Plus, commaJoin0Plus, repeat, isEmpty) to JSDoc-typed
module.f.mjs under the stage-1 TypeScript-to-mjs migration.
proof.f.ts and testlib.f.ts stay TypeScript: proof.f.ts has a runtime
dependency on testlib.f.ts, which is a shared test-fixture library
(not itself a proof/module file) still authored in TypeScript.

Updates the nine other dependents (bnf/token_symbol, bnf/descent
module+proof, bnf/ll1 module+proof, bnf/data module+proof,
djs/tokenizer, fsc/json.f.ts, fsc/bnf.f.ts) to import runtime values
from module.f.mjs and types from types.ts, splitting mixed
import statements where needed. Fixes eleven stale doc references.
