+++ fjs/bnf/data/module.f.ts
@@ -14,10 +14,12 @@ import { stringToCodePointList } from '../../text/utf16/module.f.mjs'
 import { map, toArray } from '../../types/list/module.f.mjs'
 import {
     oneEncode,
-    type DataRule,
-    type Rule as FRule,
-    type Sequence as FSequence,
-} from '../module.f.ts'
+} from '../module.f.mjs'
+import type {
+    DataRule,
+    Rule as FRule,
+    Sequence as FSequence,
+} from '../types.ts'
 import { definedEntries } from '../../types/object/module.f.mjs'
 import type { StringMap } from '../../types/object/types.ts'
