+++ fjs/bnf/data/proof.f.ts
@@ -1,7 +1,7 @@
 import { stringify } from '../../media/json/module.f.ts'
 import { identity } from '../../types/function/module.f.mjs'
 import { sort } from '../../types/object/module.f.mjs'
-import { oneEncode, option, range, rangeDecode, repeat0Plus, set } from '../module.f.ts'
+import { oneEncode, option, range, rangeDecode, repeat0Plus, set } from '../module.f.mjs'
 import { classic, deterministic } from '../testlib.f.ts'
 import { emptyTagMap, type RuleSet, toData } from './module.f.ts'
 import { assertEq } from '../../asserts/module.f.mjs'
