+++ fjs/bnf/ll1/proof.f.ts
@@ -1,6 +1,6 @@
 import { stringToCodePointList } from '../../text/utf16/module.f.mjs'
 import { toArray } from '../../types/list/module.f.mjs'
-import { commaJoin0Plus, option, range, repeat0Plus, set } from '../module.f.ts'
+import { commaJoin0Plus, option, range, repeat0Plus, set } from '../module.f.mjs'
 import { deterministic } from '../testlib.f.ts'
 import { type RuleSet, toData } from '../data/module.f.ts'
 import { dispatchMap, type MatchResult, parser, parserRuleSet } from './module.f.ts'
