+++ fjs/bnf/todo/207.md
@@ -41,7 +41,7 @@ hypotheses actually tested in §4.
 
 ### 1. Background: rule kinds and the AST
 
-A functional grammar rule (`fjs/bnf/module.f.ts`) is one of four shapes, behind
+A functional grammar rule (`fjs/bnf/module.f.mjs`) is one of four shapes, behind
 an optional lazy thunk:
 
 ```ts
@@ -244,7 +244,7 @@ Two integration strategies:
 
 A core design principle of this repository is a clean three-way split:
 
-- **one grammar definition** — the functional `Rule` form (`fjs/bnf/module.f.ts`);
+- **one grammar definition** — the functional `Rule` form (`fjs/bnf/module.f.mjs`);
 - **one serializable form** — the function-free data form (`fjs/bnf/data`);
 - **many parsers** — LL(1) (`parserRuleSet`), recursive descent
   (`descentParser`), and potentially LR(1), LR(k), GLR, … each a separate
