+++ fjs/bnf/todo/terminal-range-representation.md
@@ -97,5 +97,5 @@ assuming that only `TerminalRange` changes.
   decision.
 - [Separate alphabet-specific BNF helpers](./unicode-rules.md) — keeps the range
   representation independent from Unicode-specific syntax.
-- [`fjs/bnf/module.f.ts`](../module.f.ts) — current 24-bit packed range encoding and
+- [`fjs/bnf/module.f.mjs`](../module.f.ts) — current 24-bit packed range encoding and
   rule representation.
