+++ fjs/fsc/json.f.ts
@@ -1,4 +1,5 @@
-import { join0Plus, rangeEncode, range, remove, repeat0Plus, type Rule, set, type TerminalRange, option } from '../bnf/module.f.ts'
+import { join0Plus, rangeEncode, range, remove, repeat0Plus, set, option } from '../bnf/module.f.mjs'
+import type { Rule, TerminalRange } from '../bnf/types.ts'
 
 // space
