+++ fjs/text/todo/190.md
@@ -21,7 +21,7 @@ const { fromCharCode } = String
 export const listToString: (input: List<U16>) => string
     = fn(map(String.fromCharCode)) /* … */
 
-// fjs/bnf/module.f.ts:71
+// fjs/bnf/module.f.mjs:71
 const { fromCodePoint } = String
 ```
